+++ src/Controller/Api/Stations/Files/BatchAction.php
@@ -388,6 +388,7 @@ final class BatchAction
 
             if (!isset($queuedMediaUpdates[$mediaId])) {
                 $message = new Message\ReprocessMediaMessage();
+                $message->storage_location_id = $storageLocation->getIdRequired();
                 $message->media_id = $mediaId;
                 $message->force = true;
 
@@ -400,7 +401,7 @@ final class BatchAction
 
             if (!isset($queuedNewFiles[$path])) {
                 $message = new Message\AddNewMediaMessage();
-                $message->storage_location_id = (int)$storageLocation->getId();
+                $message->storage_location_id = $storageLocation->getIdRequired();
                 $message->path = $unprocessable->getPath();
 
                 $this->messageBus->dispatch($message);